refactor(@ddcp/models,@ddcp/manager,@ddcp/selector,@ddcp/sns-to-github,@ddcp/sns-to-slack,@ddcp/synthesizer): shared models
