--- packages/lib-sns-models/.npmignore
@@ -1,3 +1,0 @@
-*
-!dist/index.js
-!dist/index.d.ts
