--- packages/lib-sns-models/CHANGELOG.md
@@ -1,120 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [1.5.1-138](https://github.com/curquhart/ddcp/compare/v1.5.0-123...v1.5.1-138) (2020-04-01)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-# [1.5.0-123](https://github.com/curquhart/ddcp/compare/v1.4.0-121...v1.5.0-123) (2020-03-31)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-# [1.4.0-121](https://github.com/curquhart/ddcp/compare/v1.3.1-115...v1.4.0-121) (2020-03-30)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-## [1.3.1-115](https://github.com/curquhart/ddcp/compare/v1.3.0-112...v1.3.1-115) (2020-03-30)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-# [1.3.0-112](https://github.com/curquhart/ddcp/compare/v1.2.1-107...v1.3.0-112) (2020-03-30)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-## [1.2.1-107](https://github.com/curquhart/ddcp/compare/v1.2.0-104...v1.2.1-107) (2020-03-29)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-# [1.2.0-104](https://github.com/curquhart/ddcp/compare/v1.1.2-102...v1.2.0-104) (2020-03-29)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-## [1.1.2-102](https://github.com/curquhart/ddcp/compare/v1.1.1-100...v1.1.2-102) (2020-03-29)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-## [1.1.1-100](https://github.com/curquhart/ddcp/compare/v1.1.0-beta.97...v1.1.1-100) (2020-03-29)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-# [1.1.0-beta.97](https://github.com/curquhart/ddcp/compare/v1.1.0-beta.94...v1.1.0-beta.97) (2020-03-29)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-# [1.1.0-beta.94](https://github.com/curquhart/ddcp/compare/v1.1.0-91...v1.1.0-beta.94) (2020-03-29)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-# [1.1.0-91](https://github.com/curquhart/ddcp/compare/v1.0.0...v1.1.0-91) (2020-03-28)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-# [1.0.0](https://github.com/curquhart/ddcp/compare/v0.0.1-beta.88...v1.0.0) (2020-03-28)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-## [0.0.1-beta.88](https://github.com/curquhart/ddcp/compare/v0.0.1-alpha.81...v0.0.1-beta.88) (2020-03-28)
-
-**Note:** Version bump only for package @ddcp/sns-models
-
-
-
-
-
-## [0.0.1-alpha.81](https://github.com/curquhart/ddcp/compare/v0.0.1-alpha.79...v0.0.1-alpha.81) (2020-03-28)
-
-**Note:** Version bump only for package @ddcp/sns-models
