--- packages/lib-sns-models/jest.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-  roots: [
-    'src'
-  ],
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  collectCoverage: true,
-  coverageReporters: ['json', 'html'],
-  coverageDirectory: 'coverage'
-};
