--- packages/lib-sns-models/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@ddcp/sns-models",
-  "version": "1.5.1-138",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "tsc",
-    "clean": "rm -rf node_modules dist",
-    "lint": "eslint --ext .ts src/",
-    "test": "jest"
-  },
-  "engines": {
-    "node": ">=10.0.0"
-  },
-  "eslintConfig": {
-    "extends": [
-      "@ddcp/eslint-config"
-    ]
-  },
-  "dependencies": {
-    "aws-lambda": "1.0.5"
-  },
-  "devDependencies": {
-    "@babel/preset-typescript": "7.9.0",
-    "@ddcp/eslint-config": "^1.5.1-138",
-    "@ddcp/typescript-config": "^1.0.0",
-    "@types/aws-lambda": "8.10.44",
-    "@types/jest": "25.1.4",
-    "@typescript-eslint/eslint-plugin": "2.21.0",
-    "@typescript-eslint/parser": "2.21.0",
-    "eslint": "6.8.0",
-    "eslint-plugin-node": "11.0.0",
-    "jest": "25.1.0",
-    "ts-jest": "25.2.1",
-    "typescript": "3.8.2"
-  }
-}
