--- packages/lib-sns-models/src/__tests__/Example.test.ts
@@ -1,5 +1,0 @@
-describe('Example test', () => {
-    it('is an example test', () => {
-        expect(true).toBeTruthy();
-    });
-});
