--- packages/lib-sns-models/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@ddcp/typescript-config",
-  "files": [
-    "src/index.ts"
-  ],
-  "compilerOptions": {
-    "outDir": "dist",
-    "declaration": true
-  }
-}
